CPP-61 Paysheet Print

+++ PayrollAPI/Models/Payroll/Temp_Payroll.cs
@@ -44,8 +44,8 @@ namespace PayrollAPI.Models.Payroll
         [Column(TypeName = "decimal(10, 2)")]
         public decimal balanceamount { get; set; }
 
-        public float epfConRate { get; set; }
-        public float taxConRate { get; set; }
+        public double epfConRate { get; set; }
+        public double taxConRate { get; set; }
 
         [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
         public DateTime createdDate { get; set; }
